Add files via upload
Most customers do not post a review rating or any comment after purchasing a product which is a challenge for any E-commerce platform to perform If a company predicts whether a customer liked/disliked a product so that they can recommend more similar and related products as well as they can decide whether or not a product should be sold at their end. This is crucial for E-commerce-based company because they need to keep track of each product of each seller, so that none of products discourage their customers to come shop with them again. Moreover, if a specific product has very few ratings and that too negative, a company must not drop the product straight away, may be many customers who found the product to be useful haven't actually rated it. Some reasons could possibly be comparing your product review with those of your competitors beforehand, gaining lots of insight about the product and saving a lot of manual data pre-processing, maintain good customer relationship with company, lend gifts, offers and deals if the company feels the customer is going to break the relation. Objective of this case study is centered around predicting customer satisfaction with a product which can be deduced after predicting the product rating a user would rate after he makes a purchase.